- Remove Order::OrderInnerStatus: we are calculating the order status from all of their products status - Update Products::Add(): add $family property - Update Products->GetFamily(): change the property checks - Update Products->SetFamily(): change the property $family to type Family - Update Seller->Fire(): check if the seller has some families under his responsibility if he have then the fire will throw exception and will exit without to fire - Update Seller->Remove(): check if the seller has some families under his responsibility if he have then the Remove will throw exception and will exit without to Remove
Family Class:
- Update Family::$responsibleSeller property to type Seller
- Add Family::LoopAll() function
- Add Family::GetSellerResponsibility(): return array with the seller responsible families
- Add Family->__toString()

+++ Classes/Seller.php
@@ -201,6 +201,10 @@ class Seller {
      * @throws \Exception
      */
     public function Fire() {
+        $responsibleFamilies = Family::GetSellerResponsibility($this);
+        if (count($responsibleFamilies) > 0)
+            throw new Exception("Unable to Fire seller {0} because he is responsible over {1} product families!", $responsibleFamilies, $this, count($responsibleFamilies));
+
         $this->sellerStatus = ESellerStatus::Fired();
         $this->update();
 
@@ -225,6 +229,10 @@ class Seller {
      * @throws \Exception
      */
     public function Remove() {
+        $responsibleFamilies = Family::GetSellerResponsibility($this);
+        if (count($responsibleFamilies) > 0)
+            throw new Exception("Unable to delete seller {0} because he is responsible over {1} product families!", $responsibleFamilies, $this, count($responsibleFamilies));
+
         $success = BugOrderSystem::GetDB()->where(self::TABLE_KEY_COLUMN, $this->id)->delete(self::TABLE_NAME, 1);
         if(!$success)
             throw new DBException("Unable to remove seller {0}", null, $this);
